feat: remove create event form step 3
